requirements.txt
plexapi==4.15.4
pyyaml==6.0.1
python-logging-loki==0.3.1